Add each scenario instance
